api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyGithub==2.1.1
GitPython==3.1.41
pinecone-client==2.2.4
torch==2.2.0
transformers==4.38.1
sentence-transformers==2.5.1
gunicorn==21.2.0
slither-analyzer==0.10.1